Rename * all refs to com.rpm renamed to info.masterfrog
