+++ pom.xml
@@ -4,15 +4,15 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.rpm.pixelcat</groupId>
+    <groupId>info.masterfrog.pixelcat</groupId>
     <artifactId>pixelcat-demo</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <dependencies>
         <dependency>
-            <groupId>com.rpm.pixelcat</groupId>
+            <groupId>info.masterfrog.pixelcat</groupId>
             <artifactId>pixelcat-engine</artifactId>
-            <version>1.0-SNAPSHOT</version>
+            <version>0.1-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.google.guava</groupId>
@@ -79,7 +79,7 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>${project.build.finalName}.lib/</classpathPrefix>
-                            <mainClass>com.rpm.pixelcat.demo.main.DemoMain</mainClass>
+                            <mainClass>info.masterfrog.pixelcat.demo.main.DemoMain</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
