+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.rpm.pixelcat.demo.main.DemoMain
+Main-Class: info.masterfrog.pixelcat.demo.main.DemoMain
