+++ src/main/java/info/masterfrog/pixelcat/demo/gameobject/builder/CatHostGenerator.java
@@ -1,22 +1,22 @@
-package com.rpm.pixelcat.demo.gameobject.builder;
+package info.masterfrog.pixelcat.demo.gameobject.builder;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
-import com.rpm.pixelcat.engine.common.printer.Printer;
-import com.rpm.pixelcat.engine.common.printer.PrinterFactory;
-import com.rpm.pixelcat.engine.exception.TransientGameException;
-import com.rpm.pixelcat.engine.hid.HIDEventEnum;
-import com.rpm.pixelcat.engine.logic.animation.AnimationFactory;
-import com.rpm.pixelcat.engine.logic.animation.AnimationSequence;
-import com.rpm.pixelcat.engine.logic.camera.Camera;
-import com.rpm.pixelcat.engine.logic.gameobject.GameObject;
-import com.rpm.pixelcat.engine.logic.gameobject.GameObjectManager;
-import com.rpm.pixelcat.engine.logic.gameobject.behavior.*;
-import com.rpm.pixelcat.engine.logic.gameobject.feature.*;
-import com.rpm.pixelcat.engine.logic.resource.Resource;
-import com.rpm.pixelcat.engine.logic.resource.ResourceFactory;
-import com.rpm.pixelcat.engine.logic.resource.SoundResource;
-import com.rpm.pixelcat.engine.logic.resource.SpriteSheet;
+import info.masterfrog.pixelcat.engine.common.printer.Printer;
+import info.masterfrog.pixelcat.engine.common.printer.PrinterFactory;
+import info.masterfrog.pixelcat.engine.exception.TransientGameException;
+import info.masterfrog.pixelcat.engine.hid.HIDEventEnum;
+import info.masterfrog.pixelcat.engine.logic.animation.AnimationFactory;
+import info.masterfrog.pixelcat.engine.logic.animation.AnimationSequence;
+import info.masterfrog.pixelcat.engine.logic.camera.Camera;
+import info.masterfrog.pixelcat.engine.logic.gameobject.GameObject;
+import info.masterfrog.pixelcat.engine.logic.gameobject.GameObjectManager;
+import info.masterfrog.pixelcat.engine.logic.gameobject.behavior.*;
+import info.masterfrog.pixelcat.engine.logic.gameobject.feature.*;
+import info.masterfrog.pixelcat.engine.logic.resource.Resource;
+import info.masterfrog.pixelcat.engine.logic.resource.ResourceFactory;
+import info.masterfrog.pixelcat.engine.logic.resource.SoundResource;
+import info.masterfrog.pixelcat.engine.logic.resource.SpriteSheet;
 
 import java.awt.*;
 import java.util.HashMap;
